DRY For SignIn and SignUp
fixes #16
